# src/main/resources/static/api-docs.yaml
openapi: 3.0.2
info:
  title: connector-administration-fachadmin-rest-api
  version: 1.5.31
  description: version of connector-administration-fachadmin-rest-api 1.5.31
...
servers:
  - url: https://selfadmin-nb-iot.int2-ram.m2m.telekom.com
    description: test developer environment (int2-ram)
  - url: https://selfadmin-nb-iot.test-ram.m2m.telekom.com
    description: test preprod environment (test-ram)
paths:
  /devices:
    summary: Devices management resource
    description: API is targeted at Customer Users and allows management of devices
    servers:
      - url: https://{fachAdminApiUrl}
        description: ''
        variables:
          fachAdminApiUrl:
            default: selfadmin-nb-iot.int2-ram.m2m.telekom.com
            description: int2-ram environment
    get:
      tags:
        - Devices
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Device'
          description: List of devices
      summary: Get list of devices
      description: Returns all devices belonging to the X-Customer-MandantID
    post:
      requestBody:
        description: Request body in JSON format
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Device'
        required: true
      tags:
        - Devices
      responses:
        '201':
          description: Device created
        '404':
          description: Device not found
      summary: Create device
      description: "Creates Device CoT + NBIoT Connector. Payload contains name, iccid, password, deviceType ID and contextDistribution = true in addition to the CoT username, password.\r\nThe Cloud of Things ID will be generated by the CoT"
    parameters:
      - examples:
          X-Custom-AuthenticationID:
            value: '2'
        name: X-Custom-AuthenticationID
        description: "Notice that in example request authentication (X-Custom-AuthenticationID: 2) is performed against CoT and that’s why it is important to have nbiot-cot-bridge subscribed by tenant: to allow credentials to be auto discoverable by NBIoT Connectors fachadmin service. Currently this sort of automatic tenant registration is recommended.\r\n\r\nNotice that in case of authentication against connector 'X-Custom-AuthenticationID: 1` one should use user with role FACHADMIN."
        schema:
          type: string
        in: header
        required: true
      - examples:
          X-Custom-MandantID:
            value: t24021684
        name: X-Custom-MandantID
        description: tenant domain or tenantId that has bridge microservice deployed
        schema:
          type: string
        in: header
        required: true
      - examples:
          X-Custom-User:
            value: QXBpdXNlcjE6QXBpdXNlcjE=
        name: X-Custom-User
        description: header uses encoded NBIoT Connector user credentials
        schema:
          type: string
        in: header
        required: true
...
  /devices/{id}:
    summary: Device management resource
    description: device management by id
    put:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Device'
        required: true
      tags:
        - Devices
      responses:
        '204':
          description: Device updated
        '404':
          description: Device not found
      summary: Update device
      description: Update device endpoint
    delete:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContextDistributionRequest'
        required: true
      tags:
        - Devices
      responses:
        '204':
          description: Device deleted
        '404':
          description: Device with id {} is not found in device repository
      summary: Delete device
      description: Deletes device by id. Payload contains contextDistribution = true/false
    parameters:
      - name: X-Custom-AuthenticationID
        description: "Notice that in example request authentication (X-Custom-AuthenticationID: 2) is performed against CoT and that’s why it is important to have nbiot-cot-bridge subscribed by tenant: to allow credentials to be auto discoverable by NBIoT Connectors fachadmin service. Currently this sort of automatic tenant registration is recommended.\r\n\r\nNotice that in case of authentication against connector 'X-Custom-AuthenticationID: 1` one should use user with role FACHADMIN."
        schema:
          type: integer
        in: header
        required: true
      - name: X-Custom-MandantID
        description: tenant domain or tenantId that has bridge microservice deployed
        schema:
          type: string
        in: header
        required: true
      - name: X-Custom-User
        description: header uses encoded NBIoT Connector user credentials
        schema:
          type: string
        in: header
        required: true
      - name: id
        description: ID of the device entity in database that will be updated
        schema:
          type: string
        in: path
        required: true
...
  /devices/{tenant}/{managedObjectId}:
    summary: Delete device
    description: Delete device by Tenant, Managed Object id and contextDistribution
    delete:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContextDistributionRequest'
        required: true
      tags:
        - Devices
      responses:
        '204':
          description: Device deleted
        '404':
          description: Tenant and/or CloudOfThingsDeviceId not found!
      summary: Delete device
      description: 'Endpoint for deleting device by tenant, managed object ID and contextDistribution '
    parameters:
      - name: managedObjectId
        description: managed object ID of the device
        schema:
          type: string
        in: path
        required: true
      - name: tenant
        description: name of the tenant to which device belongs
        schema:
          type: string
        in: path
        required: true
      - name: X-Custom-AuthenticationID
        description: "Notice that in example request authentication (X-Custom-AuthenticationID: 2) is performed against CoT and that’s why it is important to have nbiot-cot-bridge subscribed by tenant: to allow credentials to be auto discoverable by NBIoT Connectors fachadmin service. Currently this sort of automatic tenant registration is recommended.\r\n\r\nNotice that in case of authentication against connector 'X-Custom-AuthenticationID: 1` one should use user with role FACHADMIN."
        schema:
          type: integer
        in: header
        required: true
      - name: X-Custom-MandantID
        description: tenant domain or tenantId that has bridge microservice deployed
        schema:
          type: string
        in: header
        required: true
      - name: X-Custom-User
        description: header uses encoded NBIoT Connector user credentials
        schema:
          type: string
        in: header
        required: true
...
  /tenants/{tenantName}/devices:
    summary: Delete devices
    description: Delete devices by tenant
    delete:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContextDistributionRequestBody'
      tags:
        - Tenants
      responses:
        '204':
          description: Device deleted
        '404':
          description: Device is not registered yet in CoT and cannot be deleted
      summary: Delete devices
      description: Deletes device by Tenant name
    parameters:
      - name: tenantName
        schema:
          type: string
        in: path
        required: true
      - name: X-Custom-AuthenticationID
        description: "Notice that in example request authentication (X-Custom-AuthenticationID: 2) is performed against CoT and that’s why it is important to have nbiot-cot-bridge subscribed by tenant: to allow credentials to be auto discoverable by NBIoT Connectors fachadmin service. Currently this sort of automatic tenant registration is recommended.\r\nNotice that in case of authentication against connector 'X-Custom-AuthenticationID: 1` one should use user with role FACHADMIN."
        schema:
          type: integer
        in: header
        required: true
      - name: X-Custom-MandantID
        description: tenant domain or tenantId that has bridge microservice deployed
        schema:
          type: string
        in: header
        required: true
      - name: X-Custom-User
        description: header uses encoded NBIoT Connector user credentials
        schema:
          type: string
        in: header
        required: true
...
components:
  schemas:
    Device:
      description: ''
      maxProperties: 24
      minProperties: 24
      required:
        - name
        - contextDistribution
        - iccid
        - cloudOfThingsPassword
        - cloudOfThingsUsername
        - active
      type: object
      properties:
        id:
          description: 'Device ID. This ID is used by NBIoT Connector as CoT username for device created in the following format: device_<ID from csv file>'
          type: string
          example: 11111deviceid22222233333
        name:
          description: Device name
          type: string
          example: MQTTSD113000111
        blacklist:
          description: adittional flag
          type: boolean
          readOnly: true
        contextDistribution:
          description: Additional flag to save device in CoT/NBIoT Connector
          type: boolean
        iccid:
          writeOnly: false
          description: This id is used by NBIoT Connector during creation of a NBIoT Connector user for this device. This is important field for connector as it will be used during device authentication while connect to MQTT Broker
          maxLength: 15
          minLength: 15
          type: string
          readOnly: false
        cloudOfThingsPassword:
          description: Device password in CoT
          maxLength: 8
          minLength: 8
          type: string
          example: Device_123
        cloudOfThingsUsername:
          description: User name in CoT
          maxLength: 15
          minLength: 15
          type: string
        cloudOfThingsDeviceId:
          description: Device id in CoT
          type: string
        tenant:
          $ref: '#/components/schemas/Tenant'
        deviceType:
          $ref: '#/components/schemas/DeviceType'
        createdAt:
          description: device creation date
          type: string
        updatedAt:
          description: device update date
          type: string
        active:
          type: boolean
    Tenant:
      description: Tenant to be used for device onboarding is subscribed to Nbiot-cot-bridge CoT microservice. Hence without tenant subscribed to this CoT microservice there is no way for NBIoT Connector access CoT.
      required:
        - name
      type: object
      properties:
        name:
          description: Tenant name
          type: string
...
    ContextDistributionRequest:
      description: holder of boolean value - if device also needs to be deleted from CoT
      required:
        - contextDistribution
      type: object
      properties:
        contextDistribution:
          description: boolean value - if device also needs to be deleted from CoT
          type: boolean
    DeviceType:
      description: Device type
      required:
        - id
        - name
      type: object
      properties:
        id:
          description: Device type id
          type: string
        name:
          description: 'Device type name. Should be one of the either following list: MQTT, NBIOT, TOIL or the name of custom device type'
          type: string
        publishTopics:
          description: ''
          type: array
          items:
            type: string
        subscribeTopics:
          description: ''
          type: array
          items:
            type: string
        plugin:
          $ref: '#/components/schemas/Plugin'
    Plugin:
      description: 'NBIoT Connector contains a suitable plugin for device type. By default NBIoT Connector contains following plugins: SmartRestPlugin - Smart REST is CoT feature that allows to make REST API communication more light-weight by support of payload templates (has to be created in CoT first). Considered to be used mostly with MQTT devices. PrototypeHubPlugin - Plugin to support MQTT-SN devices, PrototypeHubPluginToil - Plugin to be used with TOIL devices'
      type: object
      properties:
        id:
          description: plugin id
          type: string
        name:
          description: plugin name
          type: string
        version:
          description: plugin version
          type: string
        fileId:
          description: ''
          type: string
        className:
          description: ''
          type: string
        active:
          description: ''
          type: boolean
        metaData:
          $ref: '#/components/schemas/PluginMetaData'
...
    ContextDistributionRequestBody:
      description: holder of boolean value - if device also needs to be updated in CoT
      type: object
      properties:
        contextDistribution:
          description: ''
          type: boolean
    PluginMetaData:
      description: ''
      type: object
      properties:
        uploadedBy:
          description: ''
          type: string
        uploadedAt:
          $ref: '#/components/schemas/Date'
        updatedBy:
          description: ''
          type: string
        updatedAt:
          $ref: '#/components/schemas/Date'
    Date:
      description: represents a specific instant in time, with millisecond precision
      type: string
      example: '2020-02-06T16:44:46.221+00:00'
...
  securitySchemes:
    Network:
      scheme: basic
      type: http